Move asIsFunction() to ComponentFunction as a static method.
