+++ src/main/java/flora/command/newbuilder/component/factory/RequirementComponentFactory.java
@@ -5,7 +5,7 @@ import flora.command.newbuilder.component.applier.GenericComponentApplier;
 
 import java.util.function.Predicate;
 
-import static flora.command.newbuilder.component.factory.CentralComponentFactory.asIsFunction;
+import static flora.command.newbuilder.component.ComponentFunction.asIsFunction;
 
 public class RequirementComponentFactory {
     public static <S> GenericComponentApplier<S> requires(final Predicate<S> requirement) {
